requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.21.6
packaging==24.1
pandas==1.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-surprise==1.1.4
scipy==1.7.3
six==1.16.0
soupsieve==2.6
tmdbv3api==1.9.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
